Move path resolution logic to a separate file
