+++ compiler/noirc_frontend/src/elaborator/mod.rs
@@ -47,6 +47,7 @@ use crate::{
 mod comptime;
 mod expressions;
 mod lints;
+mod path_resolution;
 mod patterns;
 mod scope;
 mod statements;
